Day 6 part 2 solved
Ok, that took some thinking and discussion in the group to realize that
I just needed to brute force the answer. Instead of trying to predict
where you could place an obstacle to loop back on your previous path,
you really have to just test placing an obstacle on every cell that you
would normally walk through without turning. This is because you have to
account for a new obstacle sending you towards obstacles you wouldn't
normally have encountered that could put you into a loop.

Took around 30 seconds to run on the given input, which is orders of
magnitude longer than any solution up to this point.