WIP: Day 11 Part 2
Naively running through the problem for 75 iterations will not work if
you want the answer in this geological era. This is my first attempt to
optomize a solution, but it's not performing well yet and I don't have
confidence that I'd get the solution in my lifetime as it runs now.

But the basic approach would be to save the evolution of each stone in
case you encounter stones with the same values, you can re-use previous
work.