Day 11 part 2 solved
Ok, so in my previous attempt I was still trying to compute n iterations
of blink for each individual number and saving the results, which was
still the same amount of work in the end.

When I debugged my code I saw that as the iterations grew, the number of
distinct stone values grew much slower (something like ~50 distinct
values for ~3000 stones). Stones never affect their neighbor, so despite
the problem explicitly saying the stones stay in the same order, that
doesn't affect the total number of stones at the end of a blink. So I
reduced the problem down to just keeping track of the count of each
stone value at each step, which finished very quickly:

Part 1: <REDACTED> (0:00:00.004030)
Part 2: <REDACTED> (0:00:00.097461)